add /thanks to build export toggle functions from separated file use relatives paths
