+++ package.json
@@ -14,7 +14,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "prebuild": "rm -rf ./dist && mkdir dist && rm -rf ./.cache && mkdir .cache",
-    "build": "parcel build src/*.html src/presentador/*/*.html",
+    "build": "parcel build src/*.html src/presentador/*/*.html src/thanks/*.html",
     "dev": "parcel src/index.html src/thanks/index.html"
   },
   "repository": {
