+++ src/script.js
@@ -1,4 +1,5 @@
 import { setPreferredLanguage, toggleLanguage } from "./i18n";
+import { toggleNav, toggleDark, handleMode } from "./toggle-functions";
 
 document.addEventListener("DOMContentLoaded", handleMode);
 document.addEventListener("DOMContentLoaded", setPreferredLanguage);
@@ -97,31 +98,3 @@ function parseBody(validator) {
   body += "form-name=contact";
   return body;
 }
-
-function toggleNav() {
-  document.querySelector("nav").classList.toggle("hide");
-  document.querySelector(".button-menu").classList.toggle("opened");
-}
-
-function toggleDark() {
-  document.querySelector("body").classList.toggle("dark-mode");
-  document.querySelector("body").classList.toggle("light-mode");
-}
-
-function handleMode() {
-  if (window.matchMedia) {
-    const darkModeMediaQuery = window.matchMedia(
-      "(prefers-color-scheme: dark)"
-    );
-    darkModeMediaQuery.addListener((e) => {
-      toggleDark();
-    });
-    if (darkModeMediaQuery.matches) {
-      document.querySelector("body").classList.toggle("dark-mode");
-    } else {
-      document.querySelector("body").classList.toggle("light-mode");
-    }
-  }
-}
-
-export { toggleNav, toggleDark, handleMode };
