+++ src/thanks/index.html
@@ -57,7 +57,7 @@
         </h4>
         <p id="thanks-pending" class="hide">
           Your payment for <span class="service-name"></span> is pending yet.
-          <a href="https://rusticit.com/#want_more">Contact us</a> for specific
+          <a href="/#want_more">Contact us</a> for specific
           information.
         </p>
         <p id="thanks-approved" class="hide">
